backend/requirements_full.txt
cryptography
Django
django_ratelimit
djangorestframework
djangorestframework_simplejwt
dropbox
google_api_python_client
google_auth_oauthlib
numpy
openai
openpyxl
paddleocr
pdf2image
Pillow
pillow_heif
protobuf
py7zr
python-dotenv
python_dateutil
qrcode
rarfile
reportlab
Requests
scikit_learn
urllib3
wagtail
